requirements.txt
streamlit
moviepy
numpy
chromadb
torch
transformers
pillow